+++ internal/bind/auto_declare.go
@@ -4,7 +4,7 @@ import (
 	"sort"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xstring"
+	"github.com/ydb-platform/ydb-go-sdk/v3/pkg/xstring"
 )
 
 type AutoDeclare struct{}