+++ table/table.go
@@ -76,6 +76,10 @@ type Client interface {
 	// be upserted and some might not.
 	BulkUpsert(ctx context.Context, table string, data BulkUpsertData, opts ...Option) error
 
+	DescribeTable(ctx context.Context, path string,
+		opts ...options.DescribeTableOption,
+	) (desc *options.Description, err error)
+
 	// ReadRows reads a batch of rows non-transactionally.
 	ReadRows(
 		ctx context.Context, path string, keys types.Value,